blog - function | fix banner post deleted image

+++ function/functions.php
@@ -241,10 +241,15 @@ function ubahPost($data)
     #unlink($path . $gambar_lama);
   #}
 
+#  if ($gambar_lama != "default.png") {
+#    unlink($path . $gambar_lama);
+#  }
+
   if ($gambar_lama != "default.png") {
-    unlink($path . $gambar_lama);
+    if($gambar_lama != $gambar && $gambar != '') {
+      unlink($path . $gambar_lama);
+    }
   }
-  
 
   $query = "UPDATE post SET
     judul = '$judul',
